This uses the same code structure as SimRunner2, but it doesn't use a dictionary since dictionaries have a constant overhead.

Benchmark on my PC:
Running 10 times:
SimRunner: 850ms
SimRunner2: 600 to 1000 ms (has a lot of variation)
SimRunner3: 570ms

Running 100 times:
SimRunner: 5.3s
SimRunner2: 5.5s to 10s
SimRunner3: 4s