@Helpsypoo 2024-02-16
Thanks for this!

I combined some parts of this with SimRunner2 to make SimRunner4, which I plan to run with for now.

The things I didn't keep from this:
- Memory pooling. This has a big impact on the speed of multi-runs, but the multi-runs are only for increasing the sample size for testing, and actual use will usually involve only one sim at a time. I'll keep MemoryPool in the Utilities folder for future use, though.
- Entities as structs. This is probably the better way to do things in this simple simulation, but I'm interested in getting used to an extreme composition-based approach where entities are just IDs. But SimRunner4 uses lists to track the links instead of dictionaries, which is much faster.